Bound the undo log, shrink its op to 20 bytes, and split app.c
The undo op carried two whole tokens, 228 bytes, for a tile paint that
needs two. Token payloads now live in a side array the log owns and each
op records where its slots start, so truncating the redo tail truncates
both. The history is capped at four full fills of the largest map and
trims its oldest quarter in one memmove past that, once per 262,144 ops
rather than once per keystroke. Zones: undo.step, undo.trim; scenarios:
build fill+undo, build fill history. PERFORMANCE.md regenerated, with the
README's summary numbers brought up to the same run.

app.c loses the : command line (app_cmd.c) and the play-mode key handler
(app_play.c); app_priv.h carries what the three share. Nothing moved
changed.

Smaller: die is _Noreturn; the map loader treats an over-long line as one
line rather than two; tokens compare by field, not memcmp; the 5-10-5
integer halving is explained; three int-then-widen multiplications and
four misplaced size_t casts tidied.

+++ src/main.c
@@ -16,7 +16,7 @@
 /* Upper bound on bytes taken from the terminal in one pass of the event
  * loop, so a flood of input can never postpone the redraw indefinitely. Far
  * larger than any burst a held key produces. */
-#define INPUT_DRAIN_MAX (64 * 1024)
+#define INPUT_DRAIN_MAX ((size_t)64 * 1024)
 
 typedef struct {
     const char *trace_path;
